+++ src/main/java/com/edcapplication/model/TestBed.java
@@ -1,16 +1,10 @@
 package com.edcapplication.model;
 
-import java.util.List;
-
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 
 @Entity
@@ -19,7 +13,7 @@ public class TestBed {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private int id;
+	private Long id;
 	
 	@Column(name= "testbed_name")
 	private String name;
@@ -28,17 +22,17 @@ public class TestBed {
 		super();
 	}
 
-	public TestBed(int id, String name) {
+	public TestBed(Long id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 	
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
