+++ src/main/java/com/edcapplication/repository/EquipmentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.edcapplication.model.Equipment;
 
 @Repository
-public interface EquipmentRepository extends CrudRepository<Equipment, Integer> {
+public interface EquipmentRepository extends CrudRepository<Equipment, Long> {
 
 }
