+++ src/main/java/com/edcapplication/repository/ProjectRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.edcapplication.model.Project;
 
 @Repository
-public interface ProjectRepository extends CrudRepository<Project, Integer> {
+public interface ProjectRepository extends CrudRepository<Project, Long> {
 
 }
